changed OwnerSystem and ContainerSystem to use enable/disable instead added/removed and added a test to check the case when the owner of another entity was deleted and we want children to be disabled if they have a store

+++ commons-gdx-artemis/src/main/java/com/gemserk/commons/artemis/systems/ContainerSystem.java
@@ -3,6 +3,7 @@ package com.gemserk.commons.artemis.systems;
 import com.artemis.Entity;
 import com.artemis.EntityProcessingSystem;
 import com.gemserk.commons.artemis.components.ContainerComponent;
+import com.gemserk.commons.artemis.components.StoreComponent;
 
 public class ContainerSystem extends EntityProcessingSystem {
 
@@ -12,12 +13,17 @@ public class ContainerSystem extends EntityProcessingSystem {
 	}
 
 	@Override
-	protected void removed(Entity e) {
+	protected void disabled(Entity e) {
 		ContainerComponent containerComponent = ContainerComponent.get(e);
 		for (int i = 0; i < containerComponent.getChildren().size(); i++) {
 			Entity child = containerComponent.getChildren().get(i);
 			// what about the stores?
-			child.delete();
+
+			StoreComponent storeComponent = StoreComponent.get(child);
+			if (storeComponent != null)
+				storeComponent.store.free(child);
+			else
+				child.delete();
 		}
 	}
